requirements.txt
pygame-ce>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.3
groq>=0.4.0